requirements.txt
anyio==4.3.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
pyTelegramBotAPI==4.17.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
telebot==0.0.5
tornado==6.1
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
